Ref: Move some functions from KeySerialization to NewDerivationKeyFactory
